feat(space): move space identity (name) to top of overview, above card status
Extracted identity strip into SpaceIdentityStrip; order is now identity → card
overview → KPI/details. Removed strip + unused vars/imports from SpaceDetailContent.
